[AIRFLOW-3476,3477] Move Kube classes out of models.py
